requirements.txt
anyio==3.7.0
asttokens==2.2.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.1.0
backcall==0.2.0
bcrypt==4.0.1
cffi==1.16.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cryptography==41.0.5
cycler==0.11.0
# databases==0.8.0
debugpy==1.6.7
decorator==5.1.1
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.0.0.post2
executing==1.2.0
fastapi==0.96.0
fastapi-users==12.1.2
fastapi-users-db-sqlalchemy==6.0.1
fastjsonschema==2.17.1
fonttools==4.40.0
greenlet==2.0.2
h11==0.14.0
httptools==0.6.1
idna==3.4
imageio==2.31.1
ipykernel==6.23.1
ipython==8.14.0
jedi==0.18.2
joblib==1.2.0
jsonschema==4.18.2
jsonschema-specifications==2023.6.1
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.4
makefun==1.15.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
nbformat==5.9.1
nest-asyncio==1.5.6
numpy==1.24.3
packaging==23.1
pandas==2.0.2
parso==0.8.3
passlib==1.7.4
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.3
plotly==5.15.0
postgres==4.0
prompt-toolkit==3.0.38
psutil==5.9.5
psycopg2==2.9.6
psycopg2-binary==2.9.6
psycopg2-pool==1.1
pure-eval==0.2.2
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.8
Pygments==2.15.1
PyJWT==2.8.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
# pywin32==306
PyYAML==6.0.1
pyzmq==25.1.0
referencing==0.29.1
rpds-py==0.8.10
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy
sqlalchemy-citext==1.8.0
SQLAlchemy-Utils==0.41.1
stack-data==0.6.2
starlette==0.27.0
tenacity==8.2.2
threadpoolctl==3.1.0
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.6.3
tzdata==2023.3
uvicorn==0.22.0
watchfiles==0.21.0
wcwidth==0.2.6
websockets==12.0
